chore(network-socket-exchange-event): adjusted client to wait for connection when issuing event, adjusted server to wait for dependencies before starting
