+++ packages/api/src/socket/client.ts
@@ -1,7 +1,6 @@
 import { io, Socket } from 'socket.io-client';
 import { IMessage, SocketEvents, SocketUsernames } from './types';
 
-
 export class SocketClient {
   _socket: Socket = null;
 
@@ -19,9 +18,26 @@ export class SocketClient {
     this._socket = io(URL, { autoConnect: true, auth });
   }
 
+  private async _emitWhenConnected(event: string, data: any) {
+    if (this._socket.connected) {
+      this._socket.emit(event, data);
+    } else {
+      await new Promise<void>(resolve => {
+        this._socket.once('connect', () => {
+          resolve();
+        });
+      });
+      this._socket.emit(event, data);
+    }
+  }
+
   // Método para enviar uma mensagem para o servidor
-  sendMessage(message: IMessage) {
-    this._socket.emit(SocketEvents.DEFAULT, message);
+  async sendMessage(message: IMessage) {
+    await this._emitWhenConnected(SocketEvents.DEFAULT, message);
+  }
+
+  async emit(event: string, data: any) {
+    await this._emitWhenConnected(event, data);
   }
 
   // Método para desconectar do servidor Socket.IO
@@ -30,6 +46,6 @@ export class SocketClient {
   }
 
   get socket() {
-    return this._socket
+    return this._socket;
   }
 }
